+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log
 
 
+## [v0.3.0] -
+
+
+### Added
+
+- command **lukso network validator setup range --from [FROM_POSITION] --to [TO_POSITION]
+- added possibility to run l16 with **lukso network init --chain l16**
+
+### Fixed
+
+- spelling of LYX
+
+## [v0.2.2] -
+
+
+### Fixed
+
+- amount in validator setup
 
 ## [v0.2.2] -